Added Jest setup script rather than having to import every time
