+++ src/frontend/components/help-view/lib/component.test.ts
@@ -1,4 +1,3 @@
-import "testing/globals.definitions";
 import HelpView from "./component";
 import {describe, expect, it} from "@jest/globals";
