+++ src/frontend/js/lib/util/domutils/lib/elementFactory.test.ts
@@ -1,4 +1,3 @@
-import 'testing/globals.definitions';
 import { createElement } from "./elementFactory";
 import { describe, it, expect } from "@jest/globals"
