+++ src/frontend/testing/setup.ts
@@ -1,0 +1,12 @@
+import { jest } from "@jest/globals";
+
+declare global {
+    interface Window {
+        $: JQueryStatic;
+        jQuery: JQueryStatic;
+        alert: (message?: any)=>void;
+    }
+}
+
+window.$ = window.jQuery = require("jquery");
+window.alert = jest.fn();
